requirements.txt
django
gunicorn
django-heroku
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
python-telegram-bot